feat: Replace full removal of verified domain check with whitelist
