+++ app/src/main/java/fe/linksheet/extension/android/ResolveInfoExt.kt
@@ -2,10 +2,6 @@ package fe.linksheet.extension.android
 
 import android.content.Context
 import android.content.pm.ResolveInfo
-import android.content.pm.verify.domain.DomainVerificationManager
-import android.content.pm.verify.domain.DomainVerificationUserState
-import android.os.Build
-import androidx.annotation.RequiresApi
 import fe.linksheet.resolver.DisplayActivityInfo
 
 fun ResolveInfo.toDisplayActivityInfo(context: Context) = DisplayActivityInfo(
@@ -30,17 +26,3 @@ fun Map<String, ResolveInfo>.toDisplayActivityInfos(
 }
 
 fun Iterable<ResolveInfo>.toPackageKeyedMap() = associateBy { it.activityInfo.packageName }
-
-@RequiresApi(Build.VERSION_CODES.S)
-fun ResolveInfo.hasVerifiedDomains(
-    manager: DomainVerificationManager,
-    linkHandlingAllowed: Boolean
-) = manager.getDomainVerificationUserState(activityInfo.packageName)?.let { state ->
-    val linkHandling =
-        if (linkHandlingAllowed) state.isLinkHandlingAllowed else !state.isLinkHandlingAllowed
-    val hasAnyVerifiedOrSelected = state.hostToStateMap.any {
-        it.value == DomainVerificationUserState.DOMAIN_STATE_VERIFIED || it.value == DomainVerificationUserState.DOMAIN_STATE_SELECTED
-    }
-
-    linkHandling && hasAnyVerifiedOrSelected
-} ?: false
